refactor(dapp): extract background request controller
